Increased modularity of the code
